+++ app/__init__.py
@@ -5,7 +5,7 @@ Flask application factory.
 
 import os
 
-__version__ = "0.1.4"
+__version__ = "0.1.5"
 
 from flask import Flask, send_from_directory